Move `InMemoryFeedStore` to its own class on production
