+++ Tests/FeedStoreChallengeTests.swift
@@ -5,43 +5,6 @@
 import XCTest
 import FeedStoreChallenge
 
-class InMemoryFeedStore: FeedStore {
-    
-    private let queue = DispatchQueue(label: "\(InMemoryFeedStore.self)Queue", attributes: .concurrent)
-    
-    private var cache: (feed: [LocalFeedImage], timestamp: Date)?
-    
-    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        queue.async(flags: .barrier) { [weak self] in
-            guard let self = self else { return }
-            
-            self.cache = nil
-            completion(nil)
-        }
-    }
-    
-    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        queue.async(flags: .barrier) { [weak self] in
-            guard let self = self else { return }
-            
-            self.cache = (feed, timestamp)
-            completion(nil)
-        }
-    }
-    
-    func retrieve(completion: @escaping RetrievalCompletion) {
-        queue.async { [weak self] in
-            guard let self = self else { return }
-            
-            guard let cache = self.cache else {
-                return completion(.empty)
-            }
-            
-            completion(.found(feed: cache.feed, timestamp: cache.timestamp))
-        }
-    }
-}
-
 class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
 	
 	func test_retrieve_deliversEmptyOnEmptyCache() {
